Feat: 자유게시판 게시글 삭제시에도 권한 확인

+++ src/controllers/postController.ts
@@ -54,9 +54,15 @@ class PostController {
     res.status(200).json(posts);
   }
 
-  async deletePostByNumber(req: Request, res: Response): Promise<void> {
+  async deletePostByNumber(req: AuthRequest, res: Response): Promise<void> {
+    // 인증된 사용자의 정보가 있는지 확인합니다.
+    if (!req.user) {
+      res.status(401).json({ message: "사용자 인증이 필요합니다." });
+      return;
+    }
     const post = await PostService.deleteByPostNumber(
       Number(req.params.postNumber),
+      req.user._id,
     );
     res.status(200).json(post);
   }
