+++ src/services/postService.ts
@@ -79,11 +79,23 @@ class PostService {
   }
 
   // 게시글 삭제 (postNumber를 기반으로)
-  async deleteByPostNumber(postNumber: number): Promise<IPost> {
-    const deletedPost = await PostRepository.deleteByPostNumber(postNumber);
-    if (!deletedPost) {
+  async deleteByPostNumber(postNumber: number, userId: string): Promise<IPost> {
+    // 게시글 조회 -> 권한 확인 -> 삭제
+
+    // 1. 게시글 조회
+    const post = await PostRepository.findByPostNumber(postNumber);
+    if (!post) {
       throw new NotFoundError("게시글을 찾을 수 없습니다.");
     }
+
+    // 2. 권한체크
+    if (post.user_id["_id"].toString() !== userId.toString()) {
+      throw new UnauthorizedError("게시글 수정 권한이 없습니다.");
+    }
+
+    // 3. 삭제
+    const deletedPost = await PostRepository.deleteByPostNumber(postNumber);
+
     return deletedPost;
   }
 }
